Removal of cache in favor of a grace timer with a scheduled cleanup event
This should fix the prior issue of a race condition leading to some
channels not being removed. This may result in temporary channels
remaining for slightly longer in some edge cases

Should this part properly fix the first issue, there will be a follow
up patch with handling for users intentionally abusing the grace period
to spam create channels.