[GuildJoinRestrict]
While I disagree with assertations that the terms `whitelist/blacklist`
are historically racist (they aren't:
https://en.wikipedia.org/wiki/Blacklisting#Origins_of_the_term )

For various reasons I do think that the
words allowlist and blocklist make for better terms.
(Less jargon based in history, more plain use.)

With that in mind, I'm merging the GuildWhitelist and GuildBlacklist
cogs, adding some functionality, and in the process taking the time to
use terminolgoy that is becoming more standard, even if I disagree
with the stated motivations many people have had for such a shift.